Finished Project

+++ prepare/Learning03/Fraction.cs
@@ -53,6 +53,6 @@ public class Fraction
     }
     public double GetDecimalValue()
     {
-        return (_top / _bottom);
+        return (double)_top / (double)_bottom;
     }
 }
